XDeepAgents/pyproject.toml
[project]
name = "xdeepagents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepagents>=0.4.12",
    "langchain-openai>=1.1.12",
    "python-dotenv>=1.2.2",
    "tavily>=1.1.0",
]
